requirement.txt
torch
numpy
pandas
scikit-learn
scipy
matplotlib
pyyaml